puppet_compiler: Add puppet version to the PCC report
We have recently migrated the puppet masters to puppet version 5.  As
such we are now migrating the PCC hosts to puppet version 5.  This CR
adds the puppet version used by pcc to the pcc report.  This is useful
to have permanently as it will allows us to test new puppet versions as
well as regressions.

Bug: T236468
Change-Id: Ia4c2c5ae13ef5d9f8e1cc898e510916a18040ef7