pick_nodes: add ability to pick nodes based on a puppet class
This CR overrides the hosts parameter to allow users to pick hosts based
on a simplified cumin puppet grammer.

Select host with the grafana class
  "Hosts: C:grafana';
Select host with the grafana role
  "Hosts: O:grafana';
Select host with the grafana::production profile
  "Hosts: P:grafana::production';

The new function tries to deduplicate the set of nodes with a specific
class, role or profile.  this is done based on the node name and the set
of tages applied to a class.  I have created a test script:

 `puppetdb1001:/home/jond/test_pcc.py`

which when given a class will output the set of nodes that would be
tested by PCC. There is some example output of this script available at
the following url:
 https://phabricator.wikimedia.org/P10710

Bug: T245288
Change-Id: I4cb43fd6a3371c69461714037055c04d98bfb715