Checkout tested patch in a branch
The CI jobs have a distracting message regarding the patch being checked
out in a detached branch:

======= 8< ======= 8< ======= 8< ======= 8< ======= 8< ======= 8< =====
  Note: checking out 'FETCH_HEAD'.

  You are in 'detached HEAD' state. You can look around, make
  experimental changes and commit them, and you can discard any commits
  you make in this state without impacting any branches by performing
  another checkout.

  If you want to create a new branch to retain commits you create, you
  may do so (now or later) by using -b with the checkout command again.
  Example:

    git checkout -b <new-branch-name>
======= 8< ======= 8< ======= 8< ======= 8< ======= 8< ======= 8< =====

This change checks out the patch in a branch named after the ref which
in turn mutes the warning entirely.

Checking out in a detached head also shows a hint to the user similar
to:

  HEAD is now at 9c4f4ab42a Some commit summary

Mimic that by showing a `git log` for the last entry.

Change-Id: Ibe0c616c5306b36fe312f594cf301438557f6027