debug_presentation: script to render HTML templates
I needed a way to render the HTML files from Jinja2 templates. This
introduces `pcc-debug-presentation` which can be used to generate HTML
pages based on a dummy change. It is loosely based upon `pcc-debug-host`
introduced by 94fa14e72. The tool can be used to get feedback regarding
Jinja2 change and can later be of good use if one wants to add another
presentation format such as json output.

To avoid cloning repositories or fetching changes, I went to use
`unitest.mock` to monkey patch code I did not want to have executed.
That breaks a few assumptions such as directories preexisting but they
are quite easy to create (albeit empties).

The script generates a host for each of the supported return code
(cancelled, core_diff, diff, error, fail, noop), generates the output
page for each of them and the "build" index file.

Example usage (stripping debug messages):

  $ pcc-debug-presentation --output /tmp/pcc-html
  INFO:debug_presentation.py:Preparing directories
  INFO:debug_presentation.py:Generating cancelled001.example.org (Cancelled)
  INFO:debug_presentation.py:Generating fail001.example.org (Both fail)
  INFO:debug_presentation.py:Generating noop001.example.org (No change)
  INFO:debug_presentation.py:Generating error001.example.org (Change fails)
  INFO:debug_presentation.py:Generating core_diff001.example.org (System changes detected)
  INFO:debug_presentation.py:Generating diff001.example.org (None system changes detected)
  Rendered files:
  /tmp/foo/output/1911/42/index.html
  /tmp/foo/output/1911/42/cancelled001.example.org/fulldiff.html
  /tmp/foo/output/1911/42/cancelled001.example.org/corediff.html
  /tmp/foo/output/1911/42/cancelled001.example.org/index.html
  /tmp/foo/output/1911/42/core_diff001.example.org/fulldiff.html
  /tmp/foo/output/1911/42/core_diff001.example.org/corediff.html
  ...

Which results in the following tree:

  /tmp/pcc-html/
  ├── 42/
  │   ├── change/
  │   │   ├── catalogs/
  │   │   └── src/
  │   ├── diffs/
  │   └── production/
  │       ├── catalogs/
  │       └── src/
└── output/
    └── 1911/
        └── 42/
            ├── cancelled001.example.org/
            │   ├── corediff.html
            │   ├── fulldiff.html
            │   └── index.html
            ├── core_diff001.example.org/
            │   ├── corediff.html
            │   ├── fulldiff.html
            │   └── index.html
            ├── diff001.example.org/
            │   ├── corediff.html
            │   ├── fulldiff.html
            │   └── index.html
            ├── error001.example.org/
            │   ├── corediff.html
            │   ├── fulldiff.html
            │   └── index.html
            ├── fail001.example.org/
            │   ├── corediff.html
            │   ├── fulldiff.html
            │   └── index.html
            ├── noop001.example.org/
            │   ├── corediff.html
            │   ├── fulldiff.html
            │   └── index.html
            └── index.html

If one reinvoke the command with the same output directory it will fail
(to prevent unexpectedly recalling a command and having the output
deleted). The deletion can be forced by additionally passing `--force`.

Change-Id: I93812357de48bbf069ef00f7bcd1cc4075e6a22d