tox: remove envdir optimizations
When test environments share the same basepython and requirements, might
want to optimize the build by having the environment directory shared
between the test environments. That was however never officially
supported by tox v3:
https://github.com/tox-dev/tox/issues/425#issuecomment-801296002
and tox v4 explicitly breaks that assumption by comparing the
environemtn based on the name of the test env (which prevents them from
being shared): https://github.com/tox-dev/tox/issues/2788

There is a plugin to filter out the test environment `name` from the
configuration when doing the comparison:
https://github.com/masenf/tox-ignore-env-name-mismatch
But I don't want to require that plugin to run tox.

As part of the switch to tox v4 (6b8a8f072416) I also had to add a hack
to allow tox v3 to provision tox v4 which was to set the environment
directory to `.pkg` for the special `.tox` environment. That felt like a
hack and is potentially fragile.

Unless a solution is found in tox v4 (rather than via a plugin), remove
the envdir configuration and the now useless tox v3 back compatibility
bit. This will cause multiple identical environments to be created
initially, they should be quite fast though since they reuse wheels from
the cache.

Bug: T348434
Change-Id: I77702f5c220deff0586424aee0d08e719ad6d384