Remove `defaultbranch=master` from .gitreview
git-review 1.25.0, released in 2015, has learned to infer the target
branch from the git tracked branch. We thus no more need to hardcode a
branch to target. When `2.x` got branched, it carried over the
`.gitreview` configuration from the `master` branch and that causes
`git-review` to send patch to `master` instead of `2.x`.

Remove the `defaultbranch` setting and replace it with `track=1` to
explicitly state we want to target the tracked branch.

While at it, explicitly disable auto rebasing via `defaultrebase=0`
which is the setting applied in all repositories.

Bug: T146293
Change-Id: Ifcf9382354c172d6a342b3e2626008f616a29acc