Fix HTML index title and make titles concises
The page title was made out of the hostname but that is not set on the
index page. Use the change id when there is no host available.

To make it easier to distingish tabs in a browser, make the title more
concise. For a change number 1911 against host srv1001.example.org the
pages titles are now:

PCC change 1911 results
PCC srv1001.example.org results

Change-Id: I282638fec36e69404e0fd592000d61ffade81e54