Ensure that build directories are cleaned up
If PCC is forcefully closed for whatever reason the cleanup method
would not be called, leaving old build directories dangling. In
extreme cases this could lead to disks filling up requirering manual
intervention.

Allow the Controller class to be used as a context manager and call
the cleanup when we're done with the controller and exists out,
regardless of why.

Bug: T348974
Change-Id: I47eaf8bacb3b1c8c543c655a714ca24182814478