fix Puppet::FileServing::Content for puppet 7
For puppet 5 catalog compilations we used the puppet master command
which somehow redirected Puppet::FileServing::Content calls internally.

However with puppet 7 we are using the puppet catalog command so puppet
tries to make a rest call to puppet:8140 to grab the file, however that
hostname resolves to the WMCS puppetserver, which is not what we want,
instead set default_file_terminus to file_server so we resolve files in
the local source code checkout.

Bug: T367547
Change-Id: Iabbdff31433d5a22f9449d9d026f5ab4f6c6d9f9